% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen3.R
\name{generate_bin_covariates}
\alias{generate_bin_covariates}
\title{Generate covariates X_i given Z_i and probabilities}
\usage{
generate_bin_covariates(df, prob)
}
\arguments{
\item{df}{data_frame generated by generate_latent_covariates}

\item{prob}{vector of probabilities of 1 for X_i. If NA is used, X_i := Z_i.}
}
\value{
a data_frame of X_i covariates
}
\description{
.. content for details ..
}
\author{
Kazuki Yoshida
}
