% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen3.R
\name{generate_bin_outcome_log_bin_treatment}
\alias{generate_bin_outcome_log_bin_treatment}
\title{Generate binary outcome given X_i and binary A_i}
\usage{
generate_bin_outcome_log_bin_treatment(df, beta0, betaX, betaA1, betaXA1)
}
\arguments{
\item{df}{data_frame containing X_i and A_i}

\item{beta0}{Intercept coefficient}

\item{betaX}{Coefficient vector for covariates X_i}

\item{betaA1}{Coefficient for A_i}

\item{betaXA1}{Interaction coefficients for covariates when A_i = 1}
}
\value{
data_frame with binary Y added
}
\description{
.. content for details ..
}
\author{
Kazuki Yoshida
}
