% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen3.R
\name{generate_bin_outcome_log_tri_treatment}
\alias{generate_bin_outcome_log_tri_treatment}
\title{Generate binary outcome given X_i and three-valued A_i}
\usage{
generate_bin_outcome_log_tri_treatment(df, beta0, betaA1, betaA2, betaX,
  betaXA1, betaXA2)
}
\arguments{
\item{df}{data_frame containing X_i and A_i}

\item{beta0}{Intercept coefficient}

\item{betaA1}{Coefficient for I(A_i = 1)}

\item{betaA2}{Coefficient for I(A_i = 2)}

\item{betaX}{Coefficient vector for covariates X_i}

\item{betaXA1}{Interaction coefficients for covariates when A_i = 1}

\item{betaXA2}{Interaction coefficients for covariates when A_i = 2}
}
\value{
data_frame with binary Y added
}
\description{
.. content for details ..
}
\author{
Kazuki Yoshida
}
