% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen3.R
\name{generate_bin_treatment}
\alias{generate_bin_treatment}
\title{Generate a binary treatment}
\usage{
generate_bin_treatment(df, alphas)
}
\arguments{
\item{df}{df generated by generate_bin_covariates}

\item{alphas}{true coefficients for the logistic treatment generating model. This vector should contain the intercept.}
}
\value{
df with treatment (A) added.
}
\description{
.. content for details ..
}
\author{
Kazuki Yoshida
}
