% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen3.R
\name{generate_data}
\alias{generate_data}
\title{Generate data based on a given specific function}
\usage{
generate_data(fun, lst_params)
}
\arguments{
\item{fun}{Data generation function such as \code{\link{generate_franklin_data}}.}

\item{lst_params}{one list of arguments that \code{fun} can take.}
}
\value{
a complete simulated data_frame
}
\description{
.. content for details ..
}
\author{
Kazuki Yoshida
}
