% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen3.R
\name{generate_data_for_all_scenarios}
\alias{generate_data_for_all_scenarios}
\title{Generate R iterations for each scenario}
\usage{
generate_data_for_all_scenarios(fun, scenarios, n_parts, R, skip, path)
}
\arguments{
\item{fun}{specific data generator function}

\item{scenarios}{a scenarios object. Each list row is a specification for a given scenario. A character column named description is expected.}

\item{n_parts}{how many subfiles to create for each scenario. Use this to ease parallelization across cluster nodes.}

\item{R}{iteration count for each subfile. The total iteration count for a given scenario is \code{parts * R}.}

\item{skip}{row numbers of the scenario object to skip data generation.}

\item{path}{path to the directory where files are generated. Include the trailing /.}
}
\value{
Use for its side effect. There is no return value. Data files are created under path.
}
\description{
Given an iteration count and a list of scenario parameters for multiple scenarios (\code{Scenarios} object), generate R datasets for each scenario. If a scenario within a \code{Scenarios} object is corrupt (not correctly a \code{ScenarioDistResNet} object), it is skipped by \code{try}.
}
\author{
Kazuki Yoshida
}
