% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen3.R
\name{generate_franklin_covariates}
\alias{generate_franklin_covariates}
\title{Generate covariates as in Franklin et al 2014}
\usage{
generate_franklin_covariates(n)
}
\arguments{
\item{n}{sample size to simulated}
}
\value{
data_frame with n rows with 10 columns (X1 through X10)
}
\description{
.. content for details ..
}
\author{
Kazuki Yoshida
}
