% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen3.R
\name{generate_franklin_data}
\alias{generate_franklin_data}
\title{Generate data by Franklin et al 2014}
\usage{
generate_franklin_data(n, alphas, beta0, betaA, betaX, betaXA)
}
\arguments{
\item{n}{Sample size}

\item{alphas}{True coefficients for the first and second treatment linear predictors. This vector should contain the intercept. alphas = c(alpha01, alphaX1, alpha02, alphaX2)}

\item{beta0}{Outcome model intercept coefficient}

\item{betaA}{Outcome model coefficient for I(A_i = 1) and I(A_i = 2)}

\item{betaX}{Outcome model coefficient vector for covariates X_i}

\item{betaXA1}{Outcome model interaction coefficients for covariates. betaXA = c(betaXA1, betaXA2)}
}
\value{
a complete simulated data_frame
}
\description{
.. content for details ..
}
\author{
Kazuki Yoshida
}
