% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen3.R
\name{generate_mvn_covariates}
\alias{generate_mvn_covariates}
\title{Generate MVN covariates Z_i}
\usage{
generate_mvn_covariates(n, mu, Sigma)
}
\arguments{
\item{n}{Sample size}

\item{mu}{Mean vector}

\item{Sigma}{Positive definite variance-covariance matrix}
}
\value{
a data_frame of Z_i MVN covariates
}
\description{
.. content for details ..
}
\author{
Kazuki Yoshida
}
