% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen3.R
\name{generate_r_times_and_save}
\alias{generate_r_times_and_save}
\title{Generate R iterations for a single scenario}
\usage{
generate_r_times_and_save(fun, scenario, R, scenario_count, part_count,
  scenario_description, path)
}
\arguments{
\item{fun}{specific data generator function}

\item{scenario}{A row of scenarios data_frame.}

\item{R}{Scalar value. Data are generated R times.}

\item{scenario_count}{Scalar value. Indicate the scenario count. Used for data file name.}

\item{part_count}{Scalar value. Indicates which subpart of the scenario this part is.}

\item{scenario_description}{A character value describing the current scenario.}

\item{path}{path to the directory where the file is generated. Include the trailing /.}
}
\value{
Use for its side effect. No return value. Save a data file in the working directory.
}
\description{
Given a parameter set for a single scenario, generate R iterations of distributed research network.
}
\author{
Kazuki Yoshida
}
