% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen3.R
\name{generate_scenario_data_frame}
\alias{generate_scenario_data_frame}
\title{Create a data_frame of list columns representing scenarios}
\usage{
generate_scenario_data_frame(lst_lst_possible_values)
}
\arguments{
\item{lst_lst_possible_values}{named list of named lists of possible values.}
}
\value{
data_frame with a \code{scenarios} class attribute holding one scenario in each row.
}
\description{
A row in the output data_frame represents a distinct scenario whereas each column represents a distinct argument to be given to the data generation function.
}
\author{
Kazuki Yoshida
}
