% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen3.R
\name{generate_sturmer_base_covariates}
\alias{generate_sturmer_base_covariates}
\title{Generate Sturmer's base covariates}
\usage{
generate_sturmer_base_covariates(n)
}
\arguments{
\item{n}{Sample size}
}
\value{
data_frame with X1-X6
}
\description{
.. content for details ..
}
\author{
Kazuki Yoshida
}
