% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen3.R
\name{generate_sturmer_unmeasured_covariates}
\alias{generate_sturmer_unmeasured_covariates}
\title{Generate Sturmer's unmeasured covariates}
\usage{
generate_sturmer_unmeasured_covariates(df, base_alphas1, base_alphas2,
  prev_params, contraindication)
}
\arguments{
\item{df}{data_frame generated by \code{\link{generate_sturmer_base_covariates}} containing X1-X6}

\item{base_alphas1}{coefficients for the first linear predictor. This vector should contain the intercept.}

\item{base_alphas2}{coefficients for the second linear predictor. This vector should contain the intercept.}

\item{prev_params}{Prevalence parameters. A vector of length 3.}

\item{contraindication}{Logical to determine whether to parametrize based on contraindication. If \code{TRUE}, the unmeasured binary variables are more common in the high PS regions (corners in the PS ternary plot). If \code{FALSE}, the unmeasured binary variables are more common in the low PS regions (edges in the PS ternary plot).}
}
\value{
data_frame with additional X7-X9, which are considered the rare unmeasured covariates in the tails of PS distribution.
}
\description{
.. content for details ..
}
\author{
Kazuki Yoshida
}
