% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen3.R
\name{generate_tri_treatment}
\alias{generate_tri_treatment}
\title{Generate a three-valued treatment}
\usage{
generate_tri_treatment(df, alphas1, alphas2)
}
\arguments{
\item{df}{df generated by generate_bin_covariates}

\item{alphas1}{true coefficients for the first linear predictor. This vector should contain the intercept.}

\item{alphas2}{true coefficients for the second linear predictor. This vector should contain the intercept.}
}
\value{
df with treatment (A) added.
}
\description{
.. content for details ..
}
\author{
Kazuki Yoshida
}
