% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen3.R
\name{print.datagen3}
\alias{print.datagen3}
\title{Print method for simulated data}
\usage{
\method{print}{datagen3}(x, ...)
}
\arguments{
\item{x}{data_frame with an additional class \code{datagen3}}

\item{...}{}
}
\value{
invisibly return a matrix object
}
\description{
.. content for details ..
}
\author{
Kazuki Yoshida
}
